move text->input
